fixed player int position to double and prepared all the rays (still with constant length)

+++ Raycasting part/raycast_utils.c
@@ -6,7 +6,7 @@
 /*   By: gleal <gleal@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/07 13:21:39 by gleal             #+#    #+#             */
-/*   Updated: 2021/03/11 21:52:32 by gleal            ###   ########.fr       */
+/*   Updated: 2021/03/12 17:54:31 by gleal            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,8 +37,8 @@ void	ft_playerinfo(t_map2d *map, t_player *joe, char **strs)
 		{
 			if (ft_strchr("NSEW", strs[(int)i][(int)j]))
 			{
-				joe->x = (j + 0.5) * map->tile_size;
-				joe->y = (i + 0.5) * map->tile_size;
+				joe->x = (double)((j + 0.5) * map->tile_size);
+				joe->y = (double)((i + 0.5) * map->tile_size);
 				joe->rotangle = spawnrotation(strs[i][j]);
 			}
 			j++;
